Started writing a custom teleportation system
The XR teleportation system is not suitable for this project. The drawback is that the XR teleportation system is activated via XR Ray Interaction. Thus, it will always be possible to teleport at any given moment. This is problematic since a user may not want to teleport but will be teleported since there is no preventing it. My custom teleportation system will solve this problem by activating the teleportation system only when a key is pressed.